# swagger-api.yaml
openapi: 3.0.3
info:
  title: Игра "5 Букв"
  description: REST API для игры "5 Букв"
  version: 1.0.11
...
servers:
  - url: http://localhost:9090
...
paths:
...
  /link/{word}:
    get:
      tags:
        - link
      summary: Загадать слово и получить ссылку "Поделиться"
      description: Возвращает ссылку с загаданным словом
      operationId: getShareLink
      parameters:
        - name: word
          in: path
          required: true
          schema:
            type: string
          example: 'лента'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkOutputDto'
        '400':
          description: Неверный запрос
        '500':
          description: Ошибка сервера
components:
  schemas:
...
    ShareLinkOutputDto:
      type: object
      properties:
        link:
          type: string
          example: "http://localhost:9090/share?secret=irasq"